requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.0
deprecation==2.1.0
fire==0.5.0
fonttools==4.34.4
idna==3.4
img2pdf==0.4.4
lxml==4.9.2
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pdf2docx==0.5.6
pikepdf==7.1.1
Pillow==9.4.0
PyMuPDF==1.21.1
pyparsing==3.0.9
PySide6==6.4.2
PySide6-Addons==6.4.2
PySide6-Essentials==6.4.2
python-docx==0.8.11
requests==2.28.2
shiboken6==6.4.2
six==1.16.0
termcolor==2.2.0
urllib3==1.26.12